Synchronize the package.json metadata
We currently have two package.json files which I want to remove but
first let's make sure that they're consistent to the extent that it's
applicable.

The license is an important point and it seems that the MIT has always
been the correct one as the LICENSE file has it too.