Merge the two package.json files
They were mostly redundant and we don't really need both of them, one is
enough and can contain all we want.

The import/type definitions paths have been adjusted as it's the root
directory that's gonna be published now, not just the dist directory.

I ran some quick tests locally and it seems to be building things fine.