Provide the error value to the orElse() callback
When I implemented orElse() in [1] I forgot about a quite important
thing: it's useful of the orElse() callback receives the error value as
a parameter so it can use it.

Rust's Result.or_else()[2] provides the value to the callback, naturally
– let's follow suit.

This will help with [3] too – a user of the library tries to do something
that's been difficult without this error value parameter.

[1] a90c931de69a ("Implement the or()/orElse() family of methods (#35)")
[2] https://doc.rust-lang.org/std/result/enum.Result.html#method.or_else
[3] https://github.com/lune-climate/ts-results-es/issues/52