Stop reusing the val property in both Ok and Err
This is a very backwards incompatible change.

Basically the problem is when we had Result<T, E> and T was actually the
same as E (or compatible enough) the following mixup was possible where
the TS compiler wouldn't say anything[1]:

    const userIdOkTest: Result<string, string> = Ok("00287b3a-12fb-455c-b3d9-e662153c91b6");
    const userIdErrTest: Result<string, string> = Err("User not found");

    await doSomethingWithUser(userIdErrTest.val); // runtime error because the compiler didn't complain.

    // More examples that are not checked for ok or err
    if (userIdOkTest.val === "test") {
        console.log("ok or err not checked but it compiles.")
    } else if (userIdErrTest.val === "test") {
        console.log("ok or err not checked but it compiles.")
    }

Eliminating the overlap between Ok and Err and providing separate
properties for each resolves the issue.

[1] https://github.com/vultix/ts-results/issues/69