Rename Some.val to value
We've done something similar to Ok/Err in [1] so I figured we could
change this one too to keep things consistent.

The change is backwards incompatible but in an explicit way – the code
using the old name will fail to compile.

We'll document the migration steps for the client code.

[1] c0c3f6c487c1 ("Stop reusing the val property in Ok and Err (#58)")