Replace boolean Option/Result flags with more Rust-like methods
First of all this is a backwards incompatible change – the migration of
the client code is straightforwards though and the compiler will
complain loudly when something's wrong.

I'd like to implement Option.isSomeAnd()[1], Result.isOkAnd()[2] and
Result.isErrAnd()[3] in the near future.

These have to be methods (because they need parameters) and it'd be very
weird if they were methods but the none/some/ok/err were just boolean
flags (without the "is" prefix at that).

The first step, therefore, is to replace the old boolean flags with a
new set of methods that are named after the original Rust methods.

The new methods are type guards[4] and should maintain good developer
experience by allowing the TS compiler to perform type narrowing
correctly.

[1] https://doc.rust-lang.org/std/option/enum.Option.html#method.is_some_and
[2] https://doc.rust-lang.org/std/result/enum.Result.html#method.is_ok_and
[3] https://doc.rust-lang.org/std/result/enum.Result.html#method.is_err_and
[4] https://www.typescriptlang.org/docs/handbook/2/narrowing.html#using-type-predicates