Remove a few no longer relevant comments
When I started working on this code I was convinced we wanted async
versions of mapOr and mapOrElse. I'm no longer particularly convinced
of that.

Why?

The value of the various composition methods in AsyncResult comes from
being able to attach multiple pieces of code to AsyncResult while
staying in the AsyncResult territory.

As soon as we leave that territory by extracting the underlying value
from AsyncResult (which mapOr() and mapOrElse() would do) we can no
longer easily keep composing.

I'm not saying the methods are definitely useless, it's just that let's
implement them if we actually find a use case.