Fix the AsyncResult.andThen type to handle known-Ok transformation
Without these overloads

    goodResult.andThen((value) => Promise.resolve(Ok(value * 2)));

was of type

    AsyncResult<uknown, string | number>

which would cause downstream problems.

Granted, this could be an edge case and the map method could be better
for this (Ok -> Ok transformation so only the value inside changes) but
still, I don't think we should be generating unknown here, hence the patch.

Resolves: https://github.com/lune-climate/ts-results-es/issues/176