Fix Result's andThen and orElse chaining
It's been reported that code like this didn't work:

    class T1 { name = "T1" as const }
    class T2 { name = "T2" as const }
    class T3 { name = "T3" as const }
    class T4 { name = "T4" as const }
    class T5 { name = "T5" as const }
    class T6 { name = "T6" as const }

    function foo1(): Result<T1, T2> {
      return 0 as any
    }
    function foo2(): Result<T3, T4> {
      return 0 as any
    }
    function foo3(): Result<T5, T6> {
      return 0 as any
    }

    const a = foo1().andThen(foo2).orElse(foo3) // error

because orElse wasn't callable on the result of the andThen call
("This expression is not callable").

As far as I can tell that was because of the different andThen and
orElse overloads with different numbers of type parameters which
complicated things for us.

The overloads were meant to narrow types if we knew

a) What types were we dealing with initially (full Result or just Ok or
   Err)
b) What did the mapper functions return (Result or always Ok or Err)

While it was nice to narrow the types like that it complicated things
for us and caused (or contributed to) the problem reported above.

The downside of this patch: if provides mappers always returning Ok or
always returning Err to andThen and orElse the result types will lose
some information compared to before.

This is technically a backwards incompatible change and I'll test it
more extensively and document the consequences of it before publishing.

Possible workarounds if someone wants to keep calling orElse and andThen
with always-Ok or always-Err mappers:

* andThen + always Ok -> map
* andThen + always Err -> not sure, maybe change what returned Ok
  earlier to not do that, depends on the situation
* orElse + always Ok -> unwrapOr
* orElse + always Err -> mapErr

[1] da3abf8e566e ("Fix for Issue#24")

Fixes: https://github.com/lune-climate/ts-results-es/issues/197